+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## [Unreleased] - 2025-05-03
 
+
+
+
 ### Added:
 - Visitor.ipdb was a major memory hog. Created a `@dataclass(slots=True)` version called VisitorInMem. Reduced memory usage from 2.65mb to 0.67mb
 - Implemented `get_field_value()` and `serialize()` helper functions for the universal `stream_delivery()` function, now that we are streaming both dict-based data, and dataclass formats. This required a small change to `stream_delivery()` as well.
 - Integrated memory reporting log into `visitor_activity_gen()` to measure and report total usage in MB and by field. Currently commented out.
+
 
 
 ## [Unreleased] - 2025.05.05
@@ -22,9 +26,12 @@
 - Pagination state resets to 1 when query parameters (filters) are changed on the visitor dashboard.
 
 
+
 ## [Unreleased] - 2025.05.06
 ### Changed:
 - Optimized the in-memory I/O stats stream by replacing dictionary-based entries with a leaner, slot-based dataclass model.
+
+
 
 ## [Unreleased] - 2025.05.08
 ### Added:
@@ -33,6 +40,8 @@
 - Modal to display GPT-4 analysis results including `risk_level`, `justification`, and `recommended_action`.
 - Conditional modal CTA button ("Start Case" or "Add to Cases") based on risk level. Need to fix this so that high-risk visitors have a case auto-created.
 
+
+
 ## [Unreleased] - 2025.05.10
 ### Added:
 - `group_by_keys()` utility function to modularize and generalize grouping. DRAGON SLAYED
@@ -40,6 +49,7 @@
 
 ### Changed:
 - Refactored `stream_delivery()` to offload grouping and pagination helpers mentioned above.
+
 
 
 ## [Unreleased] - 2025.05.12
@@ -52,6 +62,8 @@
 ### Changed:
 - Jumped into Postgres conf file to set default timezone to `'Asia/Tokyo'`. Restarted, and good to go.
 
+
+
 ## [Unreleased] - 2025.05.12
 ### Added:
 - `/flagged-visitors` and `flagged-visitors/{case_id}` endpoints. They return a thinned out list of all cases, and the details for one specific case, respectively.
@@ -61,14 +73,20 @@
 ### Changed:
 - `class VisitorsFlagged()` has been split into `class VisitorsFlaggedSummary()`, which returns an info-light list of all cases, and `class VisitorsFlagged()` which builds on top of `VisitorsFlaggedSummary()`, and is used for full case detail retrieval, as well as new case creation (still). Functional testing still successful.
 
+
+
 ## [Unreleased] -  2025.05.17
 ### Added:
 - `caseModal.js` Case Modal added with a "Suspend Account" action button, and Abuse IPDB information dropdown.
+
+
 
 ## [Unreleased] -  2025.05.19
 ### Added:
 - Implemented `vps_login_monitor.sh` script to SSH into the VPS, parse recent failed SSH login attempts, and print structured CSV output (timestamp, IP, username, message).
 - Created a `FailedLoginIntel` table to collect details on repeat-offender ip addresses, and ultimately hand off to AI for further analysis.
+
+
 
 ## [Unreleased] -  2025.05.21
 ### Added:
@@ -79,6 +97,8 @@
 - Migrated AI analysis `analyze_user()` to `services.py` and removed `visitor_analysis.py`
 - Server-level ssh config file settings altered to permit login via key only: `PasswordAuthentication no`, `ChallengeResponseAuthentication no`, `UsePAM no`.
 
+
+
 ## [Unreleased] - 2025.05.23
 ### Added:
 - `ip_analysis_gathering() -> analyze_ip_address()` AI analysis workflow added to `services.py`. Dragon note regarding asyncio.gather() scaleability added, but for current purposes, works very well as-is.
@@ -87,3 +107,17 @@
 
 ### Changed:
 - `block_ip()` is now `ip_to_blacklist()` in `utils.py`.
+
+
+
+## [Unreleased] - 2025.05.24
+### Added:
+- Implemented `get_unanalyzed_ips() -> ai_analysis_update()` crud flow. A GET to the database returns all rows missing AI analysis. They're sent over to `services.py` to be processed/analyzed by `ip_analysis_gathering()` flow. Then each row is updated with the appropriate GPT-based analysis.
+- This repeating flow has been added as an asyncio task in `async def on_startup()`, and canceled in `async def shutdown_async()`
+
+### Changed:
+- Restructured the `ActionType` enum class to streamline values with permitted values from AI analysis.
+- Appended two additional columns to the `FieldLoginIntel` table - `status` and `status_change_date`. `status` pulls from a `CurrentStatus` enum class, and displays the current status of the IP address (active, suspended, or banned). `status_change_date` is a datetime object that will be changed in tandem with a change in `status` value. This will allow for a suspended account cooldown timer in the future.
+
+### Fixed:
+- The return object from Abuse IPDB is shaped like `{"data":{<data in here>}}`, thus all `.get()` calls to the object were returning `None`, as they were not structured correctly. Altered the return value in `ipabuse_check()` and all calls are working as expected, once again.